Code Cleanup
Cleaned up some of the code. No longer have a _run_wrapper I just put that into the __call__ section of pyshell. It was small code that just formats the args to send to the run function.

I want to try an keep the run function operating as close to the original as possible.

I may move some other options out of there as well but we'll see what direction I decide to go with it.

For now no functionality changes, but we are going to tick up the version because we are now officially on PyPi as pyshell.

Updated docs